Refactor to combine types into one file
