+++ schema/mutations.js
@@ -1,4 +1,6 @@
 const graphql = require('graphql');
+const axios = require('axios');
+
 const {
   GraphQLObjectType,
   GraphQLString,
@@ -6,8 +8,7 @@ const {
   GraphQLNonNull
 } = graphql;
 
-const PetType = require('./pet_type');
-const CompanyType = require('./company_type');
+const { PetType, CompanyType } = require('./types');
 
 const mutation = new GraphQLObjectType({
   name: 'Mutation',
